+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/UiResourceImpl.kt
@@ -22,6 +22,7 @@ import de.sovity.authorityportal.api.model.UserRoleDto
 import de.sovity.authorityportal.db.jooq.enums.UserRegistrationStatus
 import de.sovity.authorityportal.web.auth.AuthUtils
 import de.sovity.authorityportal.web.auth.LoggedInUser
+import de.sovity.authorityportal.web.pages.connectormanagement.ConnectorCsvApiService
 import de.sovity.authorityportal.web.pages.connectormanagement.ConnectorManagementApiService
 import de.sovity.authorityportal.web.pages.organizationmanagement.OrganizationInfoApiService
 import de.sovity.authorityportal.web.pages.organizationmanagement.OrganizationInvitationApiService
@@ -37,6 +38,7 @@ import de.sovity.authorityportal.web.pages.usermanagement.UserUpdateApiService
 import de.sovity.authorityportal.web.pages.userregistration.UserRegistrationApiService
 import jakarta.inject.Inject
 import jakarta.transaction.Transactional
+import jakarta.ws.rs.core.HttpHeaders.CONTENT_DISPOSITION
 import jakarta.ws.rs.core.Response
 
 class UiResourceImpl : UiResource {
@@ -84,6 +86,9 @@ class UiResourceImpl : UiResource {
 
     @Inject
     lateinit var organizationUpdateApiService: OrganizationUpdateApiService
+
+    @Inject
+    lateinit var connectorCsvApiService: ConnectorCsvApiService
 
     // User info
     @Transactional
@@ -293,8 +298,27 @@ class UiResourceImpl : UiResource {
     }
 
     @Transactional
-    override fun updateOrganizationDeatils(mdsId: String, organizationDto: UpdateOrganizationDto): IdResponse {
+    override fun updateOrganizationDetails(mdsId: String, organizationDto: UpdateOrganizationDto): IdResponse {
         authUtils.requiresRole(Roles.UserRoles.AUTHORITY_ADMIN)
         return organizationUpdateApiService.updateOrganizationDetails(mdsId, organizationDto)
+    }
+
+    @Transactional
+    override fun downloadOwnOrganizationConnectorsCsv(environmentId: String): Response {
+        authUtils.requiresAnyRole(Roles.UserRoles.AUTHORITY_ADMIN, Roles.UserRoles.AUTHORITY_USER)
+        val mdsId = loggedInUser.organizationMdsId!!
+        return Response
+            .ok(connectorCsvApiService.generateConnectorCsv(mdsId, environmentId))
+            .header(CONTENT_DISPOSITION, "attachment; filename=$mdsId" + "_$environmentId.csv")
+            .build()
+    }
+
+    @Transactional
+    override fun downloadConnectorsCsv(mdsId: String, environmentId: String): Response {
+        authUtils.requiresAnyRole(Roles.UserRoles.AUTHORITY_ADMIN, Roles.UserRoles.AUTHORITY_USER)
+        return Response
+            .ok(connectorCsvApiService.generateConnectorCsv(mdsId, environmentId))
+            .header(CONTENT_DISPOSITION, "attachment; filename=$mdsId" + "_$environmentId.csv")
+            .build()
     }
 }