+++ authority-portal-backend/authority-portal-quarkus/src/test/kotlin/de/sovity/authorityportal/web/pages/connectormanagement/ConnectorCsvApiServiceTest.kt
@@ -1,0 +1,28 @@
+package de.sovity.authorityportal.web.pages.connectormanagement
+
+import io.quarkus.test.junit.QuarkusTest
+import jakarta.inject.Inject
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+
+@QuarkusTest
+class ConnectorCsvApiServiceTest {
+
+    @Inject
+    lateinit var connectorCsvApiService: ConnectorCsvApiService
+
+    @Test
+    fun generateConnectorCsv() {
+        // arrange
+        val expectedCsvContent = "\"Name\",\"Type\",\"MDS ID\",\"Environment\",\"URL\",\"Status\"\n" +
+            "\"Example Connector\",\"OWN\",\"MDSL2222CC\",\"test\",\"https://xample.test4/connector\",\"\"\n"
+
+        // act
+        val inputStream = connectorCsvApiService.generateConnectorCsv("MDSL2222CC", "test");
+        val content = inputStream.bufferedReader().use { it.readText() }
+
+        // assert
+        assertThat(inputStream).isNotNull();
+        assertThat(content).isEqualTo(expectedCsvContent)
+    }
+}